% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_assets}
\alias{build_assets}
\title{Build the Assets object for a collection}
\usage{
build_assets(thumbnail = build_thumbnail(), parquet_items = build_parquet())
}
\arguments{
\item{thumbnail}{list describing the thumbnail object(s) used in the collection}

\item{parquet_items}{list describing the parquet object(s) used in the collection}
}
\description{
Build the Assets object for a collection
}
