% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_collection}
\alias{build_collection}
\title{Build full metadata collection}
\usage{
build_collection(
  id,
  type = "Collection",
  stac_version = "1.0.0",
  license = "proprietary",
  links = build_links(),
  title = NULL,
  assets = NULL,
  extent = build_extent(),
  keywords = NULL,
  providers = NULL,
  summaries = NULL,
  description,
  item_assets = NULL,
  table_columns = NULL,
  extensions = NULL,
  publications = NULL
)
}
\arguments{
\item{id}{unique identifier for the collection}

\item{type}{always set to "Collection" for collection items}

\item{stac_version}{version of STAC to implement}

\item{license}{collection license}

\item{links}{list of related objects and their relation relationships}

\item{title}{short title for the collection}

\item{assets}{dictionary of asset objects that can be downloaded}

\item{extent}{spatial and temporal extents}

\item{keywords}{list of keywords describing the collection}

\item{providers}{all organizations capturing or processing the data or the hosting provider}

\item{summaries}{a map of property summaries, either a set of values, a range of values or a JSON schema}

\item{description}{detailed multi-line description to fully explain the collection}

\item{item_assets}{an optional dictionary of asset objects associated with the collection that can be downloaded or streamed, each with a unique key}

\item{table_columns}{column names, type, and description for data table}

\item{extensions}{a list of extensions the collection implements}

\item{publications}{list of relevant publication information for the collection}
}
\value{
a collection list object
}
\description{
Build full metadata collection
}
