% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_extent}
\alias{build_extent}
\title{Build Extent object for collection}
\usage{
build_extent(lat_min, lat_max, lon_min, lon_max, min_date, max_date)
}
\arguments{
\item{lat_min}{minimum latitude for site bounding box}

\item{lat_max}{maximum latitude for site bounding box}

\item{lon_min}{minimum longitude for site bounding box}

\item{lon_max}{maximum longitude for site bounding box}

\item{min_date}{minimum date of data}

\item{max_date}{maximum data of data}
}
\description{
Build Extent object for collection
}
