% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_group_variables.R
\name{build_group_variables}
\alias{build_group_variables}
\title{Build the Model JSONS for catalog}
\usage{
build_group_variables(
  table_schema,
  table_description,
  start_date,
  end_date,
  id_value,
  description_string,
  about_string,
  about_title,
  theme_title,
  destination_path,
  aws_download_path,
  group_var_items,
  thumbnail_link,
  thumbnail_title,
  group_var_vector,
  group_sites
)
}
\arguments{
\item{table_schema}{schema of data table, accessed through s3}

\item{table_description}{table that includes descriptions of data columns}

\item{start_date}{table that includes descriptions of data columns}

\item{end_date}{table that includes descriptions of data columns}

\item{id_value}{title for json file}

\item{description_string}{brief description to describe JSON}

\item{about_string}{string that contains the "about" asset information}

\item{about_title}{title to describe the "about" asset information}

\item{theme_title}{title for JSON (Forecast/Scores)}

\item{destination_path}{path for saving the JSON file}

\item{aws_download_path}{path for s3 data download}

\item{group_var_items}{list of variables for the group, called within generate_variable_model_items()}

\item{thumbnail_link}{link for the thumbnail image}

\item{thumbnail_title}{title for the thumbnail image}

\item{group_var_vector}{vector of variables that belong in a group}

\item{group_sites}{unique sites included in the group}
}
\description{
Build the Model JSONS for catalog
}
