% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_keywords}
\alias{build_keywords}
\title{Build keywords object for collection}
\usage{
build_keywords(keywords)
}
\arguments{
\item{keywords}{list of keywords to be describe the collection}
}
\description{
Build keywords object for collection
}
