% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_links}
\alias{build_links}
\title{Build the Links field for a collection}
\usage{
build_links(
  root,
  items = file.path(root, "items.json"),
  parent,
  self = root,
  doi = NULL,
  license = "https://creativecommons.org/publicdomain/zero/1.0/",
  about = NULL,
  example = NULL
)
}
\arguments{
\item{root}{The actual link in the format of an URL.
Relative and absolute links are both allowed.}

\item{items}{URL to any items.json}

\item{parent}{URL to parent of self}

\item{self}{URL}

\item{doi}{optional argument to be used in the cite-as link object}

\item{license}{URL of license}

\item{about}{a URL for describing the relevant organization or group for the collection}

\item{example}{URL of an example notebook}
}
\description{
Build the Links field for a collection
}
