% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build the Model JSONS for catalog}
\usage{
build_model(
  model_id,
  team_name,
  model_description,
  start_date,
  end_date,
  var_values,
  duration_names,
  site_values,
  site_table,
  model_documentation,
  destination_path,
  aws_download_path,
  collection_name,
  thumbnail_image_name,
  table_schema,
  table_description,
  full_var_df
)
}
\arguments{
\item{model_id}{list of unique model variables}

\item{team_name}{model team name}

\item{model_description}{model description provided in registration metadata}

\item{start_date}{min date of model forecasts}

\item{end_date}{max date of model forecasts}

\item{var_values}{unique variables in the model}

\item{duration_names}{unique duration values in the model}

\item{site_values}{unique sites in the model}

\item{model_documentation}{documentation to describe the model, provided as text}

\item{destination_path}{path for saving the json file}

\item{aws_download_path}{path for s3 data download}

\item{collection_name}{name of stac collection}

\item{thumbnail_image_name}{name of thumbnail image}

\item{table_schema}{schema of data table, accessed through s3}

\item{table_description}{table that includes descriptions of data columns}
}
\description{
Build the Model JSONS for catalog
}
