% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_noaa_forecast.R
\name{build_noaa_forecast}
\alias{build_noaa_forecast}
\title{Build the forecast or scores JSON files}
\usage{
build_noaa_forecast(
  table_schema,
  table_description,
  start_date,
  end_date,
  id_value,
  description_string,
  about_string,
  about_title,
  theme_title,
  destination_path,
  aws_download_path,
  link_items,
  thumbnail_link,
  thumbnail_title,
  group_sites,
  path_item
)
}
\arguments{
\item{table_schema}{schema of data table, accessed through s3}

\item{table_description}{table that includes descriptions of data columns}

\item{start_date}{table that includes descriptions of data columns}

\item{end_date}{table that includes descriptions of data columns}

\item{id_value}{title for json file}

\item{description_string}{brief description to describe JSON}

\item{about_string}{string that contains the "about" asset information}

\item{about_title}{title to describe the "about" asset information}

\item{theme_title}{title for JSON (Forecast/Scores)}

\item{destination_path}{path for saving the JSON file}

\item{aws_download_path}{path for s3 data download}

\item{link_items}{link asset items for collection (child items)}

\item{thumbnail_link}{link for the thumbnail image}

\item{thumbnail_title}{title of thumbnail image}

\item{group_sites}{sites included in the collection}

\item{path_item}{object used to fill in the download path}
}
\value{
JSON code for child items
}
\description{
Build the forecast or scores JSON files
}
