% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_parquet}
\alias{build_parquet}
\title{Build the parquet list to be used in the build_assets() function}
\usage{
build_parquet(href, title = NULL, description = NULL)
}
\arguments{
\item{href}{the actual link in the format of an URL. Relative and absolute links are both allowed}

\item{title}{optional title to be used in rendered displays of the link}

\item{description}{optional description for the parquet items}
}
\description{
Build the parquet list to be used in the build_assets() function
}
