% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_providers}
\alias{build_providers}
\title{Build providers object for the collection}
\usage{
build_providers(data_url, data_name, host_url, host_name)
}
\arguments{
\item{data_url}{URL for data provider organization}

\item{data_name}{data provider organization}

\item{host_url}{URL for data host}

\item{host_name}{name of data host}
}
\description{
Build providers object for the collection
}
