% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_publications}
\alias{build_publications}
\title{Build publication object for collection}
\usage{
build_publications(doi, citation)
}
\arguments{
\item{doi}{doi for publications that are relevant to the collection}

\item{citation}{full citation for publications that are relevant to the collection}
}
\description{
Build publication object for collection
}
