% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_thumbnail}
\alias{build_thumbnail}
\title{Build the thumbnail list to be used in the build_assets() function}
\usage{
build_thumbnail(href, title)
}
\arguments{
\item{href}{the actual link in the format of an URL. Relative and absolute links are both allowed}

\item{title}{optional title to be used in rendered displays of the link}
}
\description{
Build the thumbnail list to be used in the build_assets() function
}
