% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_authors.R
\name{generate_authors}
\alias{generate_authors}
\title{Build the Authors for asset items}
\usage{
generate_authors(metadata_table)
}
\arguments{
\item{metadata_table}{table created from model registration metadata}
}
\description{
Build the Authors for asset items
}
