% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_group_values.R
\name{generate_group_values}
\alias{generate_group_values}
\title{Generate child items for catalog group items (variable groups)}
\usage{
generate_group_values(group_values)
}
\arguments{
\item{group_values}{list of groups to organize the catalog (variable groups, themes, etc.)}
}
\description{
Generate child items for catalog group items (variable groups)
}
