% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_group_variable_items.R
\name{generate_group_variable_items}
\alias{generate_group_variable_items}
\title{Build the group variable items}
\usage{
generate_group_variable_items(variables)
}
\arguments{
\item{table_schema}{schema of data table, accessed through s3}

\item{table_description}{table that includes descriptions of data columns}

\item{start_date}{table that includes descriptions of data columns}

\item{end_date}{table that includes descriptions of data columns}

\item{description_string}{brief description to describe JSON}
}
\value{
JSON code for group variable items
}
\description{
Build the group variable items
}
