% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model_assets.R
\name{generate_model_assets}
\alias{generate_model_assets}
\title{Build the Assets objects for model JSONS}
\usage{
generate_model_assets(m_vars, m_duration, aws_path)
}
\arguments{
\item{m_vars}{list of unique model variables}

\item{m_duration}{list of unique model duration periods}

\item{aws_path}{s3 path for accessing model data}
}
\description{
Build the Assets objects for model JSONS
}
