% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model_items.R
\name{generate_model_items}
\alias{generate_model_items}
\title{Create model items}
\usage{
generate_model_items(model_list)
}
\arguments{
\item{model_list}{unique model_ids}
}
\value{
JSON code for model asset items
}
\description{
Create model items
}
