% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variable_model_items.R
\name{generate_variable_model_items}
\alias{generate_variable_model_items}
\title{Build the model items within each variable group}
\usage{
generate_variable_model_items(model_list)
}
\arguments{
\item{model_list}{list of unique models for each variable}
}
\value{
JSON code for model items
}
\description{
Build the model items within each variable group
}
