% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbox.R
\name{get_bbox}
\alias{get_bbox}
\title{Get bbox for forecast/scores and collection groups}
\usage{
get_bbox(site_metadata, sites)
}
\arguments{
\item{site_metadata}{site metadata provided as path or url}

\item{sites}{list of unique sites for the group}
}
\value{
list of site lat and lon pairs
}
\description{
Get bbox for forecast/scores and collection groups
}
