% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grouping.R
\name{get_grouping}
\alias{get_grouping}
\title{Organize the groupings for themes}
\usage{
get_grouping(inv_bucket, theme, collapse = TRUE)
}
\arguments{
\item{inv_bucket}{name of s3 bucket}

\item{theme}{theme name}
}
\description{
Organize the groupings for themes
}
