% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_coords.R
\name{get_site_coords}
\alias{get_site_coords}
\title{Get coordinates for model sites}
\usage{
get_site_coords(site_metadata, sites)
}
\arguments{
\item{site_metadata}{site metadata provided as path or url}

\item{sites}{list of unique sites for the model}
}
\value{
list of site lat and lon pairs
}
\description{
Get coordinates for model sites
}
