% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_images.R
\name{pull_images}
\alias{pull_images}
\title{Extract and save thumbnail images for catalog}
\usage{
pull_images(theme, m_id, image_name)
}
\arguments{
\item{theme}{theme to use for thumbnail}

\item{m_id}{model id to use for thumbnail}

\item{image_name}{image name}
}
\value{
JSON code for image asset
}
\description{
Extract and save thumbnail images for catalog
}
