% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{validate_json}
\alias{validate_json}
\title{Create and save JSON file for collection}
\usage{
validate_json(json_url)
}
\arguments{
\item{json_url}{url for the json file that is being validated}

\item{...}{any additional arguments can be used here}
}
\description{
Create and save JSON file for collection
}
