% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{write_stac}
\alias{write_stac}
\title{Create and save JSON file for collection}
\usage{
write_stac(collection_object, path, ...)
}
\arguments{
\item{collection_object}{full collection list created by the build_collection() function}

\item{path}{relative or absolute path for the JSON file destination}

\item{...}{any additional arguments can be used here}
}
\description{
Create and save JSON file for collection
}
