% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_links}
\alias{build_links}
\title{Build the Links field for a collection}
\usage{
build_links(href_link, cite_doi = NULL, landing_page)
}
\arguments{
\item{href_link}{The actual link in the format of an URL. Relative and absolute links are both allowed.}

\item{cite_doi}{optional argument to be used in the cite-as link object}

\item{landing_page}{a URL for describing the relevant organization or group for the collection}
}
\description{
Build the Links field for a collection
}
