% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collection.R
\name{build_table_columns_full_bucket}
\alias{build_table_columns_full_bucket}
\title{Build table_columns object for collection}
\usage{
build_table_columns_full_bucket(data_object, description_df)
}
\arguments{
\item{data_object}{parquet object used to store data}

\item{description_df}{dataframe of table column descriptions, one description for each column needed}
}
\description{
Build table_columns object for collection
}
