% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_group_variable_items.R
\name{generate_group_variable_items}
\alias{generate_group_variable_items}
\title{Build the group variable items}
\usage{
generate_group_variable_items(variables)
}
\arguments{
\item{variables}{unique variables for group}
}
\value{
JSON code for group variable items
}
\description{
Build the group variable items
}
