% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model_assets.R
\name{generate_model_assets}
\alias{generate_model_assets}
\title{Build the Assets objects for model JSONS}
\usage{
generate_model_assets(full_var_df, aws_path, model_code_link)
}
\arguments{
\item{full_var_df}{dataframe that contains variable column information (name, full name, duration, duration name)}

\item{aws_path}{s3 path for accessing model data}
}
\description{
Build the Assets objects for model JSONS
}
