% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_validate.R
\name{stac_validate}
\alias{stac_validate}
\title{stac validate}
\usage{
stac_validate(x)
}
\arguments{
\item{x}{path or url for the json file that is being validated}
}
\description{
stac validate
}
\details{
requires \code{reticulate}
}
